Log on search or filter checkbox change
